+++ src/nh99/model45.cpp
@@ -48,7 +48,7 @@ void param_init_bounded_vector_vector::allocate(
   int size = max1 - min1 + 1;
   if (size > 0)
   {
-    v = new param_init_bounded_vector[size];
+    v = new param_init_bounded_vector[static_cast<unsigned int>(size)];
     if (!v)
     {
         cerr << " error trying to allocate memory in "
